Safari-609.1.11.1/LayoutTests/css3/flexbox/overflow-keep-scrollpos.html

<!DOCTYPE html>
<html>
<head>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<style>
div.flexbox {
    display: flex;
}
div.overflow-auto {
    overflow: auto;
}
div.flex {
    flex: 1;
}
.container {
    width: 100px;
    height: 100px;
}
</style>
<script>
    function runTest() {
        document.querySelector('.overflow-auto').scrollTop = 50;
        document.querySelector('#sidebar').innerHTML = '';
        assert_equals(document.querySelector('.overflow-auto').scrollTop, 50);
    }
</script>

</head>
<body onload="test(runTest, 'Scroll position is correct')">
<div id=log></div>
<div class="container flexbox">
    <div class="flex overflow-auto flexbox">
        <div style="height: 400px"></div>
    </div>
    <div id="sidebar">foo</div>
</div>
<div id="console"></div>
</body>
</html>
